Make generator work for all exercises
Coulnd't make it work with Gulp 🦆
