+++ package.json
@@ -52,23 +52,23 @@
 		"slides:start": "mdx-deck ./slides/slides.mdx",
 		"slides:build": "mdx-deck build ./slides/slides.mdx -d ./build/slides --basepath /slides",
 		"format": "prettier --write '**/*.{js,md}'",
-		"generate": "gulp"
+		"generate": "node scripts/generate.js",
+		"generate:watch": "node scripts/generate.js --watch"
 	},
 	"devDependencies": {
 		"babel-plugin-styled-components": "^1.10.6",
 		"caniuse-lite": "^1.0.30001008",
+		"chokidar": "^3.3.0",
 		"cross-env": "^5.2.1",
 		"eslint-plugin-jsx-a11y": "^6.2.3",
-		"gulp": "^4.0.2",
-		"gulp-watch": "^5.0.1",
 		"husky": "^1.3.1",
 		"lint-staged": "^10.0.0-beta.3",
 		"mdx-deck": "^2.5.1",
+		"mkdirp": "^0.5.1",
 		"npm-run-all": "^4.1.5",
 		"prettier": "^1.18.2",
 		"raw-loader": "^3.1.0",
-		"react-styleguidist": "^9.2.0",
-		"through2": "^3.0.1"
+		"react-styleguidist": "^9.2.0"
 	},
 	"lint-staged": {
 		"*.{js}": "eslint --fix",
